VERSION 0.0.4 moved from CVS to git
Updated tino/ and bump VERSION just for git-log purpose

------------------------------------------------------------------------

VERSION 0.0.3

Fixes for wrong return values.

Returns 127 in case of child exec fails.

Returns 210 on fork fails as it ought to (instead of -1 previously)

As always, nearly not tested at all.

------------------------------------------------------------------------

VERSION 0.0.2

Minor bugfix for output in pipe case

------------------------------------------------------------------------

VERSION 0.0.1

Bugfix: Idle mode was inverted

------------------------------------------------------------------------

VERSION 0.0.0

I needed this version as a replacement for the broken variant
which comes with Debian Etch (which does not return the
return value of the forked command).

------------------------------------------------------------------------

VERSION 1.0.0

Enter announcement here.